pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helpscout-backend"
version = "1.0.0"
description = "HelpScout Backend API"
requires-python = ">=3.10.0,<3.11.0"
dependencies = [
    "Flask==2.3.3",
    "Flask-CORS==4.0.0",
    "mysql-connector-python==8.0.33",
    "python-dotenv==1.0.0",
    "bcrypt==4.1.2",
    "pytz==2024.1",
    "gunicorn==21.2.0",
    "google-generativeai==0.3.2",
    "groq==0.4.2",
]

[tool.setuptools]
py-modules = []